some more reliability immprovements
