+++ .env.example
@@ -101,3 +101,22 @@ export GE_FOLLOWS_CACHE_RETRY_COOLDOWN_SEC="60"
 # incomplete or stale) and reports it as gauges. Runs deferred, off the
 # request path.
 export GE_FOLLOWS_CACHE_SWEEP_SEC="300"
+
+########## User-History Cache ##########
+
+# Age at which cached recommendation history is served stale and refreshed in
+# the background (default 10 minutes).
+export GE_USER_HISTORY_CACHE_TTL_SEC="600"
+
+# Hard serving cutoff and Firestore expiration for cached history. Must be at
+# least GE_USER_HISTORY_CACHE_TTL_SEC (default 30 minutes).
+export GE_USER_HISTORY_CACHE_MAX_AGE_SEC="1800"
+
+# How long one instance may hold a stale-refresh lease before another retries.
+export GE_USER_HISTORY_CACHE_LEASE_SEC="30"
+
+# Budget for rebuilding stale history from Elasticsearch in the background.
+export GE_USER_HISTORY_REFRESH_TIMEOUT_SEC="10"
+
+# Quiet period after a failed refresh, preventing a retry on every request.
+export GE_USER_HISTORY_CACHE_RETRY_COOLDOWN_SEC="60"
