+++ src/app/lib/inference.py
@@ -16,7 +16,7 @@ from .feed_debug import current_recorder
 from .http_client import get_http_client
 from .request_context import get_request_id
 from .telemetry import timed
-from .user_history import fetch_user_history_features
+from .user_history_cache import fetch_user_history_features
 
 logger = logging.getLogger(__name__)
